requirements.txt
bs4
requests
mysql-connector-python
deepdiff